changed ref position and moved to matrix_props.
