+++ docs/refs.bib
@@ -801,6 +801,12 @@ primaryClass = {quant-ph}
   primaryClass={quant-ph}
 }
 
+@misc{PlanetMathCommutant, 
+   author = "{PlanetMath}", 
+   title = "Commutant", 
+   url = {https://planetmath.org/commutant} 
+}
+
 @article{Pusey_2012_On,
   title={On the reality of the quantum state},
   volume={8},
@@ -1466,12 +1472,6 @@ primaryClass = {quant-ph}
   publisher = {American Physical Society},
   doi = {10.1103/PhysRevA.77.060301},
   url = {https://link.aps.org/doi/10.1103/PhysRevA.77.060301}
-}
-
-@misc{PlanetMathCommutant, 
-   author = "{PlanetMath}", 
-   title = "Commutant", 
-   url = {https://planetmath.org/commutant} 
 }
