+++ toqito/matrix_props/__init__.py
@@ -33,3 +33,4 @@ from toqito.matrix_props.is_stochastic import is_stochastic
 from toqito.matrix_props.spark import spark
 from toqito.matrix_props.is_pseudo_unitary import is_pseudo_unitary
 from toqito.matrix_props.is_pseudo_hermitian import is_pseudo_hermitian
+from toqito.matrix_props.commutant import commutant
